pyproject.toml
[project]
name = "climate-action-concierge-copy"
version = "0.1.0"
description = "UK Climate Action Concierge with ADK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.0.0",
    "streamlit>=1.40.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
]
